consolidate tokens to erc20. Generate networks from template. Update Readme
