+++ subgraph/packages/subgraph-dufour/README.md
@@ -26,3 +26,99 @@ Although Network Registry (NR) primarily registers Safes created by the NodeStak
 
 ### NodeSafe Registry
 Although Node Safe Registry (NR) primarily registers Safes created by the NodeStakeFactory contract, it is still possible that a vanilla Safe gets registered.
+
+## Development
+1. Install
+
+```sh
+yarn install && yarn codegen
+export DUFOUR_SUBGRAPH_NAME=<subgrpah_name>
+```
+
+2. Create a project in Subgraph studio, if not done
+
+```sh
+npx graph init --studio $DUFOUR_SUBGRAPH_NAME
+npx graph auth --studio
+```
+
+3. Configure contract addresses for networks
+- Find if a network configuration file is created under `./networks/*-networks.json`. If not, create one. Here we take the `rotsee-networks.json` as an example. 
+- Update `NodeStakeFactory`, `NodeSafeRegistry` and `NetworkRegistry` contract addresses. Use the smallest "contract creation block number" of those aforementioned three contracts for ALL THE `startBlock`, including those for HOPR tokens.
+- Update the `CHANNELS_CONTRACT_ADDRESS` in `./src/constants.ts` with the "HoprChannels" contract address for the selected network
+- Once all the changes are saved, run
+
+```sh
+yarn build --network gnosis --network-file <path to the configuration file, e.g. ./networks/rotsee-networks.json>
+```
+
+4. Deploy to Subgraph Studio
+```sh
+npx graph deploy --studio $DUFOUR_SUBGRAPH_NAME
+```
+
+## Query
+
+A sample query: 
+
+```graphql
+{
+  safes(first: 5, where: {isCreatedByNodeStakeFactory: true}) {
+    id
+    balances {
+      mHoprBalance
+      wxHoprBalance
+      xHoprBalance
+    }
+    threshold
+    owners {
+      owner {
+        id
+      }
+    }
+    isCreatedByNodeStakeFactory
+    targetedModules {
+      id
+    }
+    allowances {
+      xHoprAllowance
+      wxHoprAllowance
+      mHoprAllowance
+      grantedToChannelsContract
+    }
+    addedModules {
+      module {
+        id
+      }
+    }
+    isEligibleOnNetworkRegistry
+    registeredNodesInSafeRegistry {
+      node {
+        id
+      }
+    }
+    registeredNodesInNetworkRegistry {
+      node {
+        id
+      }
+    }
+  }
+  _meta {
+    hasIndexingErrors
+    deployment
+  }
+  nodeManagementModules {
+    id
+    implementation
+    includedNodes {
+      node {
+        id
+      }
+    }
+    multiSend
+    target {
+      id
+    }
+  }
+}
+```
