--- subgraph/packages/subgraph-dufour/networks.json
@@ -1,16 +1,0 @@
-{
-  "gnosis": {
-    "HoprToken": {
-      "address": "0x66225dE86Cac02b32f34992eb3410F59DE416698"
-    },
-    "Contract": {
-      "address": "0xD4fdec44DB9D44B8f2b6d529620f9C0C7066A2c1"
-    },
-    "xHoprToken": {
-      "address": "0xD057604A14982FE8D88c5fC25Aac3267eA142a08"
-    },
-    "NodeSafeRegistry": {
-      "address": "0x3E7c4720934ff6A9FE122Cb761f36a11E9b848D9"
-    }
-  }
-}
