+++ subgraph/packages/subgraph-dufour/src/constants.ts
@@ -5,4 +5,4 @@ export const DECIMALS = "1000000000000000000"
 export const ALL_THE_SAFES_KEY = "all_the_safes"
 
 // network sepcific
-export const CHANNELS_CONTRACT_ADDRESS = "0xD057604A14982FE8D88c5fC25Aac3267eA142a08"
+export const CHANNELS_CONTRACT_ADDRESS = "0x4c3bb6470b72A749Db1bB394c024A92AB60a793e"
