+++ subgraph/packages/subgraph-dufour/src/tokens.ts
@@ -1,9 +1,9 @@
 import { Address, BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 import { TokenType } from "./types";
 import { decimalBase, decreaseBalancesTrackerForSafes, getOrInitializeAllowances, getOrInitializeBalances, increaseBalancesTrackerForSafes } from "./helper";
-import { Approval as MHoprApproval, Transfer as MHoprTransfer } from "../generated/mHoprToken/mHoprToken";
-import { Approval as WXHoprApproval, Transfer as WXHoprTransfer } from "../generated/wxHoprToken/wxHoprToken";
-import { Approval as XHoprApproval, Transfer as XHoprTransfer } from "../generated/xHoprToken/xHoprToken";
+import { Approval as MHoprApproval, Transfer as MHoprTransfer } from "../generated/mHoprToken/ERC20Token";
+import { Approval as WXHoprApproval, Transfer as WXHoprTransfer } from "../generated/wxHoprToken/ERC20Token";
+import { Approval as XHoprApproval, Transfer as XHoprTransfer } from "../generated/xHoprToken/ERC20Token";
 
 /**
  * Handler for update mHOPR token balances
